+++ test/libsolidity/syntaxTests/inheritance/override/override_implemented_and_unimplemented_with_implemented_call_via_super.sol
@@ -1,0 +1,12 @@
+contract A {
+    function f() public virtual {}
+}
+abstract contract B {
+    function f() public virtual;
+}
+contract C is A, B {
+    function f() public override(A, B) {
+        super.f(); // super should skip the unimplemented B.f() and call A.f() instead.
+    }
+}
+// ----